Conflicts:
	app/org/sagebionetworks/bridge/dao/AccountDao.java
	app/org/sagebionetworks/bridge/services/AuthenticationService.java
	app/org/sagebionetworks/bridge/services/ParticipantService.java
	test/org/sagebionetworks/bridge/services/ParticipantServiceTest.java

+++ test/org/sagebionetworks/bridge/stormpath/StormpathAccountDaoTest.java
@@ -8,13 +8,13 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
+import static org.sagebionetworks.bridge.Roles.DEVELOPER;
 import static org.sagebionetworks.bridge.Roles.TEST_USERS;
 import static org.sagebionetworks.bridge.TestConstants.TEST_STUDY_IDENTIFIER;
 
 import java.util.EnumSet;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Set;
 
 import javax.annotation.Resource;
 
@@ -32,7 +32,6 @@ import org.sagebionetworks.bridge.models.PagedResourceList;
 import org.sagebionetworks.bridge.models.accounts.Account;
 import org.sagebionetworks.bridge.models.accounts.AccountSummary;
 import org.sagebionetworks.bridge.models.accounts.Email;
-import org.sagebionetworks.bridge.models.accounts.HealthId;
 import org.sagebionetworks.bridge.models.accounts.SignIn;
 import org.sagebionetworks.bridge.models.accounts.StudyParticipant;
 import org.sagebionetworks.bridge.models.studies.Study;
@@ -78,7 +77,7 @@ public class StormpathAccountDaoTest {
     
     @Test
     public void getStudyAccounts() {
-        Iterator<Account> i = accountDao.getStudyAccounts(study);
+        Iterator<AccountSummary> i = accountDao.getStudyAccounts(study);
         
         // There's always one... the behavior of the iterator is tested separately
         assertTrue(i.hasNext());
@@ -86,7 +85,7 @@ public class StormpathAccountDaoTest {
     
     @Test
     public void getAllAccounts() {
-        Iterator<Account> i = accountDao.getAllAccounts(); 
+        Iterator<AccountSummary> i = accountDao.getAllAccounts(); 
         
         // There's always one... the behavior of the iterator is tested separately
         assertTrue(i.hasNext());
@@ -108,7 +107,9 @@ public class StormpathAccountDaoTest {
                 String email = "bridge-testing+SADT"+random+"@sagebridge.org";
                 StudyParticipant participant = new StudyParticipant.Builder().withEmail(email).withPassword(PASSWORD)
                         .withRoles(Sets.newHashSet(TEST_USERS)).build();
-                Account account = accountDao.initializeAccount(study, participant.getEmail(), participant.getPassword());
+                Account account = accountDao.constructAccount(study, participant.getEmail(), participant.getPassword());
+                accountDao.createAccount(study, account, false);
+                
                 newAccounts.add(account.getId());
             }
             // Fetch only 5 accounts. Empty search string ignored
@@ -161,17 +162,18 @@ public class StormpathAccountDaoTest {
     public void canAuthenticate() {
         String email = TestUtils.makeRandomTestEmail(StormpathAccountDaoTest.class);
         Account account = null;
-        
         try {
             StudyParticipant participant = new StudyParticipant.Builder().withEmail(email).withPassword(PASSWORD)
                         .withRoles(Sets.newHashSet(TEST_USERS)).build();
-            accountDao.initializeAccount(study, participant.getEmail(), participant.getPassword());
+            account = accountDao.constructAccount(study, participant.getEmail(), participant.getPassword());
+            accountDao.createAccount(study, account, false);
             
             account = accountDao.authenticate(study, new SignIn(email, PASSWORD));
             assertEquals(email, account.getEmail());
-            assertEquals(1, account.getRoles().size());
         } finally {
-            accountDao.deleteAccount(study, account.getId());
+            if (account != null) {
+                accountDao.deleteAccount(study, account.getId());    
+            }
         }
     }
     
@@ -182,8 +184,8 @@ public class StormpathAccountDaoTest {
         try {
             StudyParticipant participant = new StudyParticipant.Builder().withEmail(email).withPassword(PASSWORD)
                     .withRoles(Sets.newHashSet(TEST_USERS)).build();
-            account = accountDao.initializeAccount(study, participant.getEmail(), participant.getPassword());
-            
+            account = accountDao.constructAccount(study, participant.getEmail(), participant.getPassword());
+            accountDao.createAccount(study, account, false);
             try {
                 accountDao.authenticate(study, new SignIn(email, "BadPassword"));
                 fail("Should have thrown an exception");
@@ -191,7 +193,9 @@ public class StormpathAccountDaoTest {
                 assertEquals("Account not found.", e.getMessage());
             }
         } finally {
-            accountDao.deleteAccount(study, account.getId());
+            if (account != null) {
+                accountDao.deleteAccount(study, account.getId());    
+            }
         }
     }
     
@@ -211,19 +215,20 @@ public class StormpathAccountDaoTest {
                     .withSignedOn(signedOn).build();
             
             StudyParticipant participant = new StudyParticipant.Builder().withEmail(email).withPassword(PASSWORD)
-                    .withRoles(Sets.newHashSet(TEST_USERS)).build();
-            account = accountDao.initializeAccount(study, participant.getEmail(), participant.getPassword());
-            
+                    .withRoles(Sets.newHashSet(DEVELOPER, TEST_USERS)).build();
+            account = accountDao.constructAccount(study, participant.getEmail(), participant.getPassword());
+            account.setRoles(participant.getRoles());
+            accountDao.createAccount(study, account, false);
+
             assertNull(account.getFirstName()); // defaults are not visible
             assertNull(account.getLastName());
             account.setEmail(email);
             account.setAttribute("phone", "123-456-7890");
-            account.setHealthId("abc");
             account.getConsentSignatureHistory(subpop.getGuid()).add(sig);
             account.setAttribute("attribute_one", "value of attribute one");
             
             // Update Account
-            accountDao.updateAccount(study, account);
+            accountDao.updateAccount(account);
             
             // Retrieve account with ID
             Account newAccount = accountDao.getAccount(study, account.getId());
@@ -231,26 +236,16 @@ public class StormpathAccountDaoTest {
 
             // Verify that you can get the health code using the email. We still need this for MailChimp.
             String healthCode = accountDao.getHealthCodeForEmail(study, email);
-            HealthId healthId = healthCodeService.getMapping(account.getHealthId());
-            assertEquals(healthCode, healthId.getCode());
+            assertEquals(healthCode, account.getHealthCode());
             
-            // Test adding and removing some groups. This gets into verifying and avoiding saving the underlying
-            // Stormpath account. There are 4 cases to consider: (1) adding groups, (2) removing groups, (3) account in
-            // groups unchanged, (4) account not in groups unchanged. To test this, we always leave the account in
-            // TEST_USERS and never add it to ADMIN.
-            Set<Roles> roles = Sets.newHashSet(newAccount.getRoles());
-            roles.addAll(EnumSet.of(Roles.DEVELOPER, Roles.RESEARCHER, Roles.WORKER));
-            
-            newAccount.setRoles(roles);
-            accountDao.updateAccount(study, newAccount);
+            newAccount.setRoles(EnumSet.of(Roles.DEVELOPER, Roles.RESEARCHER, Roles.WORKER));
+            accountDao.updateAccount(newAccount);
 
             newAccount = accountDao.getAccount(study, account.getId());
-            assertEquals(4, newAccount.getRoles().size());
-            assertTrue(newAccount.getRoles().containsAll(EnumSet.of(Roles.DEVELOPER, Roles.RESEARCHER,
-                    Roles.TEST_USERS, Roles.WORKER)));
+            assertEquals(3, newAccount.getRoles().size());
 
             newAccount.setRoles(EnumSet.of(Roles.TEST_USERS));
-            accountDao.updateAccount(study, newAccount);
+            accountDao.updateAccount(newAccount);
 
             newAccount = accountDao.getAccount(study, account.getId());
             assertEquals(1, newAccount.getRoles().size());
@@ -259,16 +254,18 @@ public class StormpathAccountDaoTest {
             // finally, test the name
             newAccount.setFirstName("Test");
             newAccount.setLastName("Tester");
-            accountDao.updateAccount(study, newAccount);
+            accountDao.updateAccount(newAccount);
             
             newAccount = accountDao.getAccount(study, newAccount.getId());
             assertEquals("Test", newAccount.getFirstName()); // name is now visible
             assertEquals("Tester", newAccount.getLastName());
             
         } finally {
-            accountDao.deleteAccount(study, account.getId());
-            account = accountDao.getAccount(study, account.getId());
-            assertNull(account);
+            if (account != null) {
+                accountDao.deleteAccount(study, account.getId());
+                account = accountDao.getAccount(study, account.getId());
+                assertNull(account);
+            }
         }
     }
 
@@ -278,17 +275,19 @@ public class StormpathAccountDaoTest {
         Account account = null;
         try {
             StudyParticipant participant = new StudyParticipant.Builder().withEmail(email)
-                .withPassword("P@ssword`1").build();
-            account = accountDao.initializeAccount(study, participant.getEmail(), participant.getPassword());
+                .withPassword(PASSWORD).build();
+            account = accountDao.constructAccount(study, participant.getEmail(), participant.getPassword());
+            accountDao.createAccount(study, account, false);
             
             // Great... now we should be able to get a healthCode
             String healthCode = accountDao.getHealthCodeForEmail(study, email);
             assertNotNull(healthCode);
             
-            HealthId healthId = healthCodeService.getMapping(account.getHealthId());
-            assertEquals(healthCode, healthId.getCode());
+            assertEquals(healthCode, account.getHealthCode());
         } finally {
-            accountDao.deleteAccount(study, account.getId());
+            if (account != null) {
+                accountDao.deleteAccount(study, account.getId());    
+            }
         }
     }
     
@@ -299,12 +298,16 @@ public class StormpathAccountDaoTest {
         
         Account account = null;
         try {
-            account = accountDao.initializeAccount(study, participant.getEmail(), participant.getPassword());
+            account = accountDao.constructAccount(study, participant.getEmail(), participant.getPassword());
+            accountDao.createAccount(study, account, false);
+            assertNotNull(account.getId());
             
             Email emailObj = new Email(study.getStudyIdentifier(), participant.getEmail());
             accountDao.resendEmailVerificationToken(study.getStudyIdentifier(), emailObj); // now send email
         } finally {
-            accountDao.deleteAccount(study, account.getId());
+            if (account != null) {
+                accountDao.deleteAccount(study, account.getId());    
+            }
         }
     }
     
@@ -337,11 +340,12 @@ public class StormpathAccountDaoTest {
         StudyParticipant participant = new StudyParticipant.Builder().withEmail(email).withPassword(PASSWORD).build();
         Account account = null;
         try {
-            account = accountDao.initializeAccount(study, participant.getEmail(), participant.getPassword());
+            account = accountDao.constructAccount(study, participant.getEmail(), participant.getPassword());
+            accountDao.createAccount(study, account, false);
             
             account.getConsentSignatureHistory(subpop1.getGuid()).add(sig1);
             account.getConsentSignatureHistory(subpop2.getGuid()).add(sig2);
-            accountDao.updateAccount(study, account);
+            accountDao.updateAccount(account);
             
             account = accountDao.getAccount(study, account.getId());
             
@@ -356,7 +360,9 @@ public class StormpathAccountDaoTest {
             assertEquals(sig2, account.getActiveConsentSignature(subpop2.getGuid()));
 
         } finally {
-            accountDao.deleteAccount(study, account.getId());
+            if (account != null) {
+                accountDao.deleteAccount(study, account.getId());    
+            }
         }
     }
     
@@ -366,14 +372,13 @@ public class StormpathAccountDaoTest {
         assertNull(newAccount.getLastName());
         assertEquals(account.getEmail(), newAccount.getEmail());
         assertEquals(account.getAttribute("phone"), newAccount.getAttribute("phone"));
-        assertEquals(account.getHealthId(), newAccount.getHealthId());
+        assertEquals(account.getRoles(), newAccount.getRoles());
+        assertEquals(account.getHealthCode(), newAccount.getHealthCode());
         assertEquals(account.getActiveConsentSignature(subpop.getGuid()), 
                 newAccount.getActiveConsentSignature(subpop.getGuid()));
         assertEquals(account.getActiveConsentSignature(subpop.getGuid()).getSignedOn(), 
                 newAccount.getActiveConsentSignature(subpop.getGuid()).getSignedOn());
         assertEquals(signedOn, newAccount.getActiveConsentSignature(subpop.getGuid()).getSignedOn());
-        assertEquals(1, newAccount.getRoles().size());
-        assertEquals(account.getRoles().iterator().next(), newAccount.getRoles().iterator().next());
         assertEquals("value of attribute one", account.getAttribute("attribute_one"));
         assertNull(account.getAttribute("attribute_two"));
     }
